+++ src/panelWebView/components/SponsorMsg.tsx
@@ -2,9 +2,15 @@ import * as React from 'react';
 import { SPONSOR_LINK } from '../../constants/Links';
 import { HeartIcon } from './Icons/HeartIcon';
 
-export interface ISponsorMsgProps {}
+export interface ISponsorMsgProps {
+  isBacker: boolean | undefined;
+}
 
-const SponsorMsg: React.FunctionComponent<ISponsorMsgProps> = (props: React.PropsWithChildren<ISponsorMsgProps>) => {
+const SponsorMsg: React.FunctionComponent<ISponsorMsgProps> = ({ isBacker }: React.PropsWithChildren<ISponsorMsgProps>) => {
+  if (isBacker) {
+    return null;
+  }
+
   return (
     <p className={`sponsor`}>
       <a href={SPONSOR_LINK} title="Support Front Matter">